Refactor BanditTweaks features and add follower toggle
